fix(input): scale MSX mouse movement in host window coordinates
Since 396a7b0fa the host-to-MSX mouse scale was derived from
OutputSurface::getPhysicalSize(), which is expressed in physical
pixels. SDL reports mouse motion in window coordinates ('points'),
and on macOS openMSX creates its window with SDL_WINDOW_ALLOW_HIGHDPI,
so on a retina display those differ by the backing scale factor. The
MSX pointer moved at half the speed of the host pointer. Measured with
scale_factor 4: a 1280x960 point window has a 2560x1920 drawable, so
the scale was 8 points per MSX pixel where it should be 4.

Two smaller problems in the same expression, and these do affect all
platforms:
- it was an integer division, so in full screen the ratio got
  truncated. At 1080p with scale_factor 2 that gave 4 instead of 4.5;
- it used one scale for both axes, while 'horizontal_stretch' means
  MSX pixels are not drawn square. At the default of 280 the X axis
  needs to be about 14% faster than the Y axis to track the host
  pointer.

Add VisibleSurface::getMsxPixelSize(), which returns the on-screen
size of a single MSX pixel per axis, in the same unit SDL uses for
mouse coordinates. It converts the viewport from physical pixels to
points via the window size, and takes 'full_stretch' and
'horizontal_stretch' into account. Mouse now uses that, keeping the
sub-pixel remainder in a float so slow movements are still not lost.

Note that this scales for equal distance on screen, which is not the
same as equal distance in MSX pixels when horizontal_stretch is not
320: the MSX pointer then crosses fewer MSX pixels per centimetre
horizontally than vertically.

Trackball and Paddle keep their fixed SCALE of 2: they never used the
display size, and neither is a 1:1 pointer device.

+++ src/input/Mouse.hh
@@ -57,7 +57,7 @@ private:
 	int phase;
 	int xRel = 0, yRel = 0;               // latched X/Y values, these are returned to the MSX
 	int curXRel = 0, curYRel = 0;         // running X/Y values, already scaled down
-	int fractionalX = 0, fractionalY = 0; // running X/Y values, not yet scaled down
+	float fractionalX = 0.0f, fractionalY = 0.0f; // running X/Y values, not yet scaled down
 	uint8_t status = JOY_BUTTONA | JOY_BUTTONB;
 	bool mouseMode = true;
 };
